Cleanup from self-review
